// src/service-api/generated/models/authUserDtoRole.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type AuthUserDtoRole = typeof AuthUserDtoRole[keyof typeof AuthUserDtoRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AuthUserDtoRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
} as const;